+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## v5.0.2
 
 * made to support Cassandra 5.0.2
-* intra-cassandra span transfers work and will be done using binary codec
+* intra-cassandra span transfers work (fixed #2) and will be done using binary codec
 * got rid of CloserThread
+* added more regexes, and improved them (fixed #3 and #4)
 
 ## v4.1.1